feat: create main.py and have setup.* use it instead
