+++ setup.cfg
@@ -1,2 +1,65 @@
+[metadata]
+name = ufw
+version = 0.37.0-rc.1
+author = Jamie Strandboge
+author_email = jamie@strandboge.com
+url = https://github.com/jdstrand/ufw
+description = Application for managing a Linux firewall
+long_description = file: README.md
+long_description_content_type = text/markdown
+license = GPL v3
+classifiers =
+    License :: OSI Approved :: GNU General Public License v3 (GPLv3)
+    Programming Language :: Python :: 3
+    Operating System :: OS Independent
+
+[options]
+packages = ufw
+package_dir =
+    ufw = src
+python_requires >= 3.8
+
+[options.entry_points]
+console_scripts =
+    ufw = ufw:ufw
+
+# compatible with black
+[flake8]
+max-line-length = 88
+# https://github.com/PyCQA/flake8-bugbear#opinionated-warnings (B*, below)
+# C	complexity checks
+# E	pycodestyle errors
+# F	pyflakes fatals
+# W	pycodestyle warnings
+# B	Bugbear checks (default opinionated warnings, see controversial below)
+# B901	Use of 'return x' in a generator is invalid
+# B902	Invalid first argument for method (use self for inst and cls for class)
+# B903	Enforce use of collections.namedtuple (or typing.NamedTuple)
+# B950	Line too long, but allow 10% more. Should also disable pycodestyle E501
+select = C,E,F,W,B,B901,B950
+ignore =
+    # The default ignore list:
+    E203,W503,W504,E501
+    # Common definitions:
+    # D203: 1 blank line required before class docstring
+    # F401: Module imported but unused
+    # E123: closing bracket does not match indentation of opening bracket's line
+    # E203: whitespace before ':'
+    # W503: line break before binary operator (not a pep8 issue, should be ignored)
+    # W504: line break after binary operator
+    # E501: line too long (using B950 from flake8-bugbear)
+    # E402: module level import not at top of file
+    # E129: visually indented line with same indent as next logical line
+    # E222: multiple spaces after operator
+    # E231: missing whitespace after ','
+exclude =
+    .git,
+    __pycache__,
+    build,
+    staging,
+    dist,
+    *.pyc,
+    *.egg-info,
+
 [pycodestyle]
 ignore = E121,E124,E126,E127,E128,E201,E202,E226,E241,E261,E265,E501,E502
