+++ src/__init__.py
@@ -1,0 +1,6 @@
+from .main import (
+    main_ufw,
+)
+
+# for setuptools
+ufw = main_ufw
